+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/dictionary-service-factory.ts
@@ -18,6 +18,17 @@ export class DictionaryServiceFactory {
             otherwise, the service will attempt to figure out the root item for the current JSS App using GraphQL and app name.
             rootItemId: '{GUID}'
           */
+          /*
+            GraphQL endpoint may reach its rate limit with the amount of Layout and Dictionary requests it receives and throw a rate limit error.
+            GraphQL Dictionary and Layout Services can handle rate limit errors from server and attempt a retry on requests.
+            For this, specify the number of retries the GraphQL client will attempt. 
+            It will only try the request once by default.
+            retries: 'number' 
+          */
+          retries:
+            (process.env.GRAPH_QL_SERVICE_RETRIES &&
+              parseInt(process.env.GRAPH_QL_SERVICE_RETRIES, 10)) ||
+            0,
         })
       : new RestDictionaryService({
           apiHost: config.sitecoreApiHost,